test(refactor): move ContainerLogsDumper to test-util
to make it available in backup store IT tests
