+++ qa/integration-tests/src/test/java/io/camunda/zeebe/it/queryapi/QueryApiIT.java
@@ -20,9 +20,9 @@ import io.camunda.zeebe.it.queryapi.util.TestAuthorizationListener;
 import io.camunda.zeebe.it.queryapi.util.TestAuthorizationServerInterceptor;
 import io.camunda.zeebe.model.bpmn.Bpmn;
 import io.camunda.zeebe.model.bpmn.BpmnModelInstance;
-import io.camunda.zeebe.qa.util.testcontainers.ContainerLogsDumper;
 import io.camunda.zeebe.qa.util.testcontainers.ZeebeTestContainerDefaults;
 import io.camunda.zeebe.test.util.grpc.CloseAwareListener;
+import io.camunda.zeebe.test.util.testcontainers.ContainerLogsDumper;
 import io.grpc.StatusRuntimeException;
 import io.zeebe.containers.ZeebeContainer;
 import java.io.File;
