+++ test-util/src/main/java/io/camunda/zeebe/test/util/testcontainers/ContainerLogsDumper.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.qa.util.testcontainers;
+package io.camunda.zeebe.test.util.testcontainers;
 
 import java.util.Map;
 import java.util.function.Supplier;
